2018.03.12
修改bug：
1.删除管理员失败（索引错误）
2.查询组织委员班级信息失败（session设置错误）
3.学生信息管理界面查看个人工时窗体title错误
4.添加各个删除按钮的判断和删除确认
5.修改活动工时表的日期格式

+++ ExperimentTen/Infor-Data/Data.aspx.cs
@@ -135,13 +135,19 @@ namespace WHMS.Infor_Data
 
         protected void btnDelete_Click3(object sender, EventArgs e)
         {
+            if (gridExample.SelectedRowIndex < 0)
+            {
+                Alert.Show("请选择一项进行删除", "警告", MessageBoxIcon.Warning);
+            }
+            else
+            {
+                string id = gridExample.SelectedRow.Values[0].ToString();//选中行的第一列为ID
 
-            string id = gridExample.SelectedRow.Values[0].ToString();//选中行的第一列为ID
-
-            string sqlStr = "delete from Working_hours where SySe= '" + id + " '";
-            Common.ExecuteSql(sqlStr);
-            this.BindGrid3();
-            Alert.ShowInTop("删除成功", "信息", MessageBoxIcon.Information);
+                string sqlStr = "delete from Working_hours where SySe= '" + id + " '";
+                Common.ExecuteSql(sqlStr);
+                this.BindGrid3();
+                Alert.ShowInTop("删除成功", "信息", MessageBoxIcon.Information);
+            }
         }
 
         protected void gridExample_Sort(object sender, GridSortEventArgs e)
